requirements.txt
autopep8==1.5.4
Django==2.2.16
django-filter==2.3.0
djangorestframework==3.11.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
